provider: New module used to expose extension points for core services
Introducing the concept of providers: co-processes that talk with the editor
through the remote API and provide implementation for one or more core
services.

There's a `provider_register` function that has an API wrapper, this function
can be used by channels that want to self-register as a service provider.

Some old builtin vim features will be re-implemented as providers. The
`provider_feature_available` function is used to check if a provider
implementing a certain feature is available(It will be called by the `has`
vimscript function to check for features in a vim-compatible way)

This implements the provider module without exposing any extension points, which
will be done in future commits.