provider: Add support for clipboard registers.
This reimplements the '+'/'*' clipboard registers(both are aliases to the same
underlying register, no separate register for the X11 selection) on top of the
provider infrastructure.

It also adds new boolean option: 'unnamedclip', which has the same effect of
setting 'clipboard' to 'unnamed/unnamedplus' in vim

The 'clipboard' option was not reused because all values(except 'unnamedplus')
seem to be useless for Neovim, and the code to parse the option was relatively
big. The option remains for vim compatibility but it's silently ignored.