process: Ensure out/err streams are only closed after reading all data
It is possible that child processes will exit without libuv reading all data out
in the OS buffer, and if this happens the data will be lost. Use a timer to
ensure libuv has a chance to read everything, after which the streams are
force-closed to deal with processes that exit without closing the connection.